+++ syscall/samples/sched/sched_helpers.c
@@ -4,7 +4,6 @@
 #endif
 #include <stdio.h>
 #include <stdlib.h>
-#include <syscall.h>
 #include <unistd.h>
 #include <sys/types.h>
 #include <assert.h>